refactor de nombres de las clases
