--- src/main/java/p1/Pepperoni.java
@@ -1,7 +1,0 @@
-package main.java.p1;
-
-public class Pepperoni extends Pizza {
-	Pepperoni() {
-		super(PizzaTypes.PEPPERONI);
-	}
-}
