+++ src/main/java/p1/PepperoniPizza.java
@@ -1,0 +1,7 @@
+package main.java.p1;
+
+public class PepperoniPizza extends Pizza {
+	PepperoniPizza() {
+		super(PizzaTypes.PEPPERONI);
+	}
+}
